Consider libreadline/libedit only in interactive mode
There's not much point in running the entire terminal mode session
on top of libreadline or libedit, unless it is clear the session
is really running interactively on a terminal.

Thus, only use the libreadline/libedit version if stdin points to
a terminal. Otherwise, resort to the simple version without readline
support. That ensures everything works well when running it within
a pipeline.