Clarify aliases for terminal commands
Whilest q is a proper alias for quit (explicitly mentioned in the command
table) this is not strictly true for d, w and r: these only work because
at this time there are no other commands starting with d, w or r than
dump, write or read. Future extensions to the terminal may change this.