Suppress overzealous compiler warnings
Some compilers are overzealous when the unary minus operates on an
unsigned number, and issue a warning. In fact, -u is a perfectly valid
unsigned for any unsigned u, and always has been from K&R C to C++20.
-u is the same as ~u + 1. Unsigned numbers form a modulo group, where -u
is the unique unsigned number that when added to u results in 0.

It is true that neither u nor -u can be negative (because no unsigned
number is) but that should not get a compiler to issue a warning about
the completely legitimate use of -u.