Suppress invalid compiler warning
A warning when building under NetBSD 9.3 (gcc 7.5.0) complains about
a comparison of the form lmax <= lmin+MAX_PAD, which is mistaken to
be of the type (X + c) >= X despite lmin and lmax not necessarily
having the same value.

This PR tries to suppress this warning about potential signed overflow by
making lmin, lmax and llen unsigned (size_t).