Include signature in prodsig memory
Data sheets include the signature in the production signature row
(prodsig/sigrow memory) while .atdf excludes the 3 signature bytes.

For example the AVR16EB14 datasheet lets sigrow start at 0x1080
with a size of 128 while the atdf file specifies the start as
0x1083 with a size of 125.

The former makes more sense.

This commit also sets page_size to the prodsig size if it is a power of 2
otherwise it is set to 1.