Review LED support
The RDY LED is set once the programmer is initialised and switched
off when AVRDUDE exits. During reading, writing or erasing the target
the PGM LED flashes with around 2.5 Hz, whilst the VFY LED comes on
during -U verification of the uploaded contents. Errors are indicated
with the ERR LED.

Assuming AVRDUDE got to the point where LEDs are accessible and the RDY
LED was switched on then, on exit, AVRDUDE will leave the LEDs in the
following states:

| PGM | VFY | ERR | Semantics                                        |
| --- | --- | --- | ------------------------------------------------ |
| off | off | off | OK: all tasks done w/o errors                    |
| off | off | on  | Some error not related to read/write/erase       |
| on  | off | on  | Read/write/erase error                           |
| off | on  | on  | Verification error but no read/write/erase error |
| on  | on  | on  | Read/write/erase error and verification error    |

Other combinations should not show after exit.