Review write_byte() functions ...
adding code that allows read-only memories to appear written successfully
provided the data to be written is already in the memory. This is useful
so that "writing" a backup to read-only memories works
  - Without generating an error
  - Without actually trying to write to the part

Trying to write a read-only memory with a byte that is not there will
cause a write error. This is useful if for example one tries to write a
backup of a part to another part with different signature. If the
signature is written back first, this constitutes a "check we are writing
to the same type of part".