Add memory sram at the expense of data
The memory data was introduced for the sole purpose to provide an
nvm_data_offset for Microchip programmers. As an address offset it would
tell programmers to read/write in data space. It so turned out that this
nvm_data_offset always was 0x1000000 for PDI and UPDI parts.

This commit removes this artificial data memory and treats the data offset
in the code as what it currently is: a constant.

At the same time the commit introduces an sram memory with offset and size
for virtually all parts known to AVRDUDE.