Workaround for a quite strange UnicodeDecodeError
Not really understood, but with level MSG_TRACE2, on a Curiosity Nano
board, I get two instances of unicode decode errors within the trace
messages. Trying to debug them did not yield any useful result, but
catching the exception, and enforcing the type to be "str" appear to
be able to continue.

Maybe an internal bug of my Python interpreter, don't know.