Remove gcc warning when printf format is ""
This warning only occurs when checking the msg_...() functions for having
the correct parameters. The way I do that is by replacing the avrdude.h
file with one that #defines the msg_... functions to printf for which the
compiler knows which paramaters of which type are expected given that the
format is a string constant. That check, however, produces false alarms
which are not helpful.

This commit switches these unhelpful warnings from gcc off.