Move serprog enable logic to pgm->initialize()
This commits moves the logic to enable the serprog programmer to be
ready for programming to the initalize() function as it seems to be the
counterpart of disable() and is called directly before programming
starts. This allows to open() a serprog programmer without blocking the
SPI bus as 0x15 S_PIN_STATE and 0x18 S_CS_MODE is only called in
initalize before the actual programming starts.