configure.ac: Guard macros with m4_pattern_forbid
Guard m4 macro calls with m4_pattern_forbid for all
macros which do not come with Autoconf/Automake.

All other macros might not be present on the system,
and it is better to m4_pattern_forbid them and have
autoreconf fail, than to have autoreconf generate a
malformed broken configure script which then always
fails with a weird error message.