CMakeLists.txt: allow disabling Python support
In some cases, even if Swig is found and Python3 is found, it may not
be desirable to build Python support in avrdude, so this commit adds
an ENABLE_PYTHON_SUPPORT option to be able to explicitly disable using
Python support (unfortunately CMake doesn't allow passing arguments
that would prevent it from finding Swig/Python 3 if available).

To preserve existing behavior, this option defaults to enabled (ON).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>